Reduce 3px movement threshold to 1px.
The 3px threshold was added to support animations, but didn't work well
for that use case; the transform exclusion is a better approach.

The max move distance factor also helps to avoid penalizing trivial
movements.  A 1px movement (even of a large element) will have a
miniscule impact on the score, if nothing else moves in the frame.

Bug: 1003893
Change-Id: Ice2a8a3a79cf1c3c30fc4cff037eca94b5ca6954