[Trusted Types] De-flake trusted-types-eval-reporting test.
The WPT test trusted-types-eval-reporting was flaky, usually failing on the
first run, but succeeding on the second. This was a real (and potentially
serious) bug, where the CSP logic would rely on the V8 eval callback being
called, but other parts would blanket-allow eval (and hence suppress the
callback) because they didn't check for 'trusted types' condition.

Also, minor cleanups for the test.

Bug: 739170
Change-Id: I00699b115f0be474c2b6d231dcc8b884868248b7