[css-contain] Avoid marking container chain for overflow recalc
If we have strict containment we don't need to recompute overflow
in the ancestors, as nothing would change outside of
the contained element.

This patch modifies
LayoutObject::MarkContainerChainForOverflowRecalcIfNeeded()
so we stop marking containers for overflow recalc if we reach
an element with "contain: strict".

This improves performance on the test cases attached to the bug.
Those tests were only modifying an element that has an ancestor
with strict containment in a very big DOM.
However LayoutBlock::RecalcChildLayoutOverflow() was called
all the elements in the DOM when it was not actually needed.
With this change we prevent that to happen, so it's only called
for the modified element.

BUG=970224
TEST=external/wpt/css/css-contain/contain-strict-overflow-001.html

Change-Id: I7a891e9ecab626c25ae2129cbe435864eed640d5