storage_setitem.html crash reproduction.
The following change minimizes the test case for a CRASH when running the
webstorage/storage_setitem.html test.

```sh
./wpt run --webdriver-binary=path/to/chromedriver --test-type=testharness android_webview webstorage/storage_setitem.html
Running 1 tests in web-platform-tests

  ▶ CRASH [expected OK] /webstorage/storage_setitem.html

Ran 1 tests finished in 2.9 seconds.
  • 0 ran as expected. 0 tests skipped.
  • 1 tests crashed unexpectedly
```

But when run with:

```c++
interesting_strs = ["A"];
```

wpt doesn't crash, but instead hangs as so:

```sh
./wpt run --webdriver-binary=path/to/chromedriver --test-type=testharness android_webview webstorage/storage_setitem.html
Running 1 tests in web-platform-tests

  [1/1] No tests running.
```

Not sure what is crashing, and attempts to enable logging via the
`--log-wptreport` or `--log-raw` options don't produce any results. Running
with the `--webdriver-arg='--verbose'` doesn't produce any additional output
to stdout (or adb logcat) and running with the `--debugger` doesn't enter the
debugger.